inline more stuff into routes
