+++ app/routes/characters.tsx
@@ -1,5 +1,1 @@
-import { CharacterSheet } from "~/features/characters/character-sheet"
-
-export default function CharactersPage() {
-  return <CharacterSheet />
-}
+export { default } from "./characters.$id"
